backend/pyproject.toml
[tool.poetry]
name = "ai-assistant-packaging"
version = "0.1.0"
description = ""
authors = ["Gerard Church <gerardchurch133@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.44.0"
flask = "^3.0.3"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.0"
pdfminer-six = "^20240706"
chromadb = "0.4.24"
langchain-chroma = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
